API-1: making sure that only the owner can delete

+++ app/Http/Controllers/Question/DeleteController.php
@@ -10,6 +10,7 @@ class DeleteController extends Controller
 {
     public function __invoke(Question $question): Response
     {
+        $this->authorize('forceDelete', $question);
         $question->forceDelete();
 
         return response()->noContent();
